Extract the NEAR API type definitions.
