+++ util/near/go.mod
@@ -1,0 +1,7 @@
+module github.com/auroraisnear/aurora-relayer/util/near
+
+go 1.17
+
+require github.com/creachadair/jrpc2 v0.30.1
+
+require golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
